[MODS][ShowResources] Add shortcut for show/hide
