+++ src/app/core/mods/showResources/showResources.ts
@@ -1,18 +1,23 @@
 import { Logger } from "app/core/electron/logger.helper";
+import { ShortcutsHelper } from "app/core/helpers/shortcuts.helper";
 import { Mod } from "../mod";
+import { Resources, Resource, iconIdByTypeId } from "./resources";
 
 export class ShowResources extends Mod {
+    private shortcutsHelper: ShortcutsHelper;
+    
     private data: Map<number, Resources> = new Map();
     private elemIdToTypeid: Map<number, number> = new Map();
     private resourceTotal: number = 0;
     private ressourcesToSkip: number[] = [-1, 16, 105, 106, 120, 128, 168, 208];
 
     private resourcesBox: HTMLDivElement;
+    private enabled: boolean = true;
 
     startMod(): void {
         this.params = this.settings.option.vip.general;
 
-        if(/*this.params*/true) {
+        if(this.params.show_resources) {
             Logger.info('- Enabled ShowRessources');
 
             let resourcesBoxCss = document.createElement('style');
@@ -49,6 +54,9 @@ export class ShowResources extends Mod {
 
             this.wGame.document.getElementsByTagName('head')[0].appendChild(resourcesBoxCss);
 
+            this.shortcutsHelper = new ShortcutsHelper(this.wGame);
+            this.shortcutsHelper.bind(this.params.show_resources_shortcut, () => this.toggle() );
+
             this.on(this.wGame.dofus.connectionManager, 'MapComplementaryInformationsDataMessage', (e: any) => this.onMapComplementaryInfos(e.interactiveElements, e.statedElements));
             this.on(this.wGame.dofus.connectionManager, 'StatedElementUpdatedMessage', ({statedElement}) => this.onStatedElementUpdated(statedElement));
             this.on(this.wGame.dofus.connectionManager, 'GameFightStartingMessage', () => this.clearHtml());
@@ -78,7 +86,7 @@ export class ShowResources extends Mod {
             this.elemIdToTypeid.set(i.elementId, i.elementTypeId);
         });
 
-        this.create();
+        if (this.enabled) this.create();
     }
 
     private onStatedElementUpdated(statedElement: any) {
@@ -93,7 +101,7 @@ export class ShowResources extends Mod {
             } catch {
                 console.error('Unabled to update resource with typeId = ' + typeId);
             }
-            this.create();
+            if (this.enabled) this.create();
         }, 500);
     }
 
@@ -119,6 +127,12 @@ export class ShowResources extends Mod {
         }
     }
 
+    private toggle() {
+        this.enabled = !this.enabled;
+        if (!this.enabled) this.clearHtml();
+        else this.create();
+    }
+
     private clearHtml() {
         if (this.resourcesBox && this.resourcesBox.parentElement) this.resourcesBox.parentElement.removeChild(this.resourcesBox);
     }
@@ -133,110 +147,4 @@ export class ShowResources extends Mod {
         super.reset();
         this.clear();
     }  
-}
-
-export class Resources {
-    private name: string;
-    private typeId: number;
-    private resources: {elemId: number, state: State}[] = [];
-
-    constructor(resource: Resource) {
-        this.name = resource.name;
-        this.typeId = resource.elementTypeId;
-        this.addOrUpdateResource(resource);
-    }
-
-    public addOrUpdateResource(resource: Resource) {
-        if (this.resources.length > 0 && this.resources.find(r => r.elemId == resource.elementId) != undefined) {
-            this.resources.find(r => r.elemId == resource.elementId).state = resource.elementState;
-        } else {
-            this.resources.push({elemId: resource.elementId, state: resource.elementState});
-        }
-    }
-
-    public getName(): string {
-        return this.name;
-    }
-
-    public getCount(): string {
-        return this.getCountOfRessourcesCanBeUsed() + '/' + this.getTotalCount();
-    }
-
-    public getIcon(): string {
-        const iconId = iconIdByTypeId[this.typeId];
-        return (iconId != undefined) ? 'https://dofustouch.cdn.ankama.com/assets/2.34.8_kbu_6h45kmUJaqYJSzE(uwaos..pYYKs/gfx/items/' + iconId + '.png' : '';
-    }
-
-    public getTotalCount():number {
-        return this.resources.length;
-    }
-
-    public getCountOfRessourcesCanBeUsed(): number {
-        return this.resources.filter(r => r.state == State.canUse).length;
-    }
-}
-
-export interface Resource {
-    elementId: number;
-    elementTypeId: number;
-    name: string;
-    elementState: State;
-}
-
-export enum State { canUse, cantUse, inUse }
-
-export const iconIdByTypeId: any = {
-    // General
-    84: '15026',    // eau
-    // Paysan
-    38: '34009',    // blé
-    39: '34080',    // houblon
-    42: '34122',    // lin
-    43: '34082',    // orge
-    44: '34082',    // seigle
-    45: '34154',    // avoine
-    46: '34121',    // chanvre
-    47: '34083',    // malt
-    111: '34010',   // riz
-    134: '34552',   // frostiz
-    // Bucheron
-    1: '38017',     // frêne
-    8: '38092',     // chêne
-    28: '38094',    // if
-    29: '38073',    // ebène
-    30: '38140',    // orme
-    31: '38153',    // erable
-    32: '38138',    // charme
-    33: '38086',    // châtaignier
-    34: '38095',    // noyer
-    35: '38139',    // merisier
-    98: '38482',    // bombu
-    101: '38481',   // oliviolet
-    108: '38673',   // bambou
-    109: '38672',   // bambou sombre
-    110: '38671',   // bambou sacré
-    121: '38001',   // kaliptus
-    133: '38677',   // trembe
-    // Alchimiste
-    //42: '35117',    // lin
-    //46: '35156',    // chanvre
-    61: '35191',    // edelweiss
-    66: '36052',    // menthe sauvages
-    67: '36067',    // trèfle à 5 feuilles
-    68: '35192',    // orchidée freyesque
-    112: '58067',   // pandouille
-    131: '35635',   // perce-neige
-    // Mineur
-    17: '39024',    // fer
-    24: '39028',    // argent
-    25: '39022',    // or
-    26: '39076',    // pierre de beauxite
-    37: '39077',    // pierre de kobalte
-    52: '39078',    // etain
-    53: '39108',    // pierre cuivrée
-    54: '39397',    // manganèse
-    55: '39109',    // bronze
-    113: '39110',   // dolomite
-    114: '39111',   // silicate
-    135: '39112',   // obsidienne
 }
